web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy
pillow
torch
torchvision
requests
